+++ packages/rs-dpp/src/data_contract/get_property_definition_by_path.rs
@@ -8,8 +8,8 @@ use crate::errors::ProtocolError;
 use crate::util::{json_schema::JsonSchemaExt, json_value::JsonValueExt};
 
 lazy_static! {
-    static ref BASE_DOCUMENT_SCHEMA: JsonValue =
-        serde_json::from_str(include_str!("../../schema/document/documentBase.json")).unwrap();
+    static ref EXTENDED_DOCUMENT_SCHEMA: JsonValue =
+        serde_json::from_str(include_str!("../../schema/document/documentExtended.json")).unwrap();
 }
 // Get user property definition
 pub fn get_property_definition_by_path<'a>(
@@ -18,7 +18,7 @@ pub fn get_property_definition_by_path<'a>(
 ) -> Result<&'a JsonValue, ProtocolError> {
     // Return system properties schema
     if path.starts_with('$') {
-        return Ok(BASE_DOCUMENT_SCHEMA.get_value(&format!("properties.{}", path))?);
+        return Ok(EXTENDED_DOCUMENT_SCHEMA.get_value(&format!("properties.{}", path))?);
     }
 
     let mut path_components = path.split('.');